dev/ocr_service/requirements.txt
aio-pika>=9.0.0
asyncpg>=0.29.0
torch>=2.0.0
transformers>=4.45.0,<5.0.0
Pillow>=10.0.0
pymupdf>=1.23.0
timm>=1.0.0
python-dotenv>=1.0.0
